# server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gem Auto Rentals API
  description: |
    RESTful API for Gem Auto Rentals - a car rental management platform.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Roles
    - **CUSTOMER**: Regular users who can book vehicles
    - **SUPPORT**: Staff who can view and manage bookings
    - **MANAGER**: Staff with additional fleet management access
    - **ADMIN**: Full system access

    ## Rate Limiting
    API requests are rate limited to 100 requests per 15 minutes per IP address.
  version: 1.0.0
  contact:
    name: Gem Auto Rentals Support
    email: support@gemautorentals.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.gemautorentals.com/api
    description: Production server
...
paths:
...
  /bookings:
    get:
      tags: [Bookings]
      summary: List bookings
      description: Get user's bookings (or all bookings for staff)
      operationId: listBookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/BookingStatus'
        - name: userId
          in: query
          schema:
            type: string
          description: Filter by user ID (staff only)
        - name: vehicleId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Bookings retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Bookings]
      summary: Create booking
      description: Create a new vehicle booking
      operationId: createBooking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
            example:
              vehicleId: clm1234567890
              startDate: "2024-03-15"
              endDate: "2024-03-18"
              pickupLocation: Main Office
              dropoffLocation: Airport
              extras:
                insurance: true
                gps: true
                childSeat: false
                additionalDriver: false
              notes: Please have the car ready by 9 AM
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Booking'
        '400':
          description: Invalid dates or vehicle unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      description: Get paginated customer list (staff only)
      operationId: listCustomers
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by email, first name, or last name
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Customers retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Unauthorized
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Not found
  schemas:
...
    VehicleCategory:
      type: string
      enum: [ECONOMY, STANDARD, PREMIUM, LUXURY, SUV, VAN]
...
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, ACTIVE, COMPLETED, CANCELLED]
    PaymentStatus:
      type: string
      enum: [PENDING, PROCESSING, SUCCEEDED, FAILED, REFUNDED, PARTIALLY_REFUNDED]
    UserRole:
      type: string
      enum: [CUSTOMER, SUPPORT, MANAGER, ADMIN]

    # ===== Common Schemas =====
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
    BookingExtras:
      type: object
      properties:
        insurance:
          type: boolean
        gps:
          type: boolean
        childSeat:
          type: boolean
        additionalDriver:
          type: boolean
    Booking:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        vehicleId:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/BookingStatus'
        pickupLocation:
          type: string
        dropoffLocation:
          type: string
        dailyRate:
          type: number
        totalAmount:
          type: number
        extras:
          $ref: '#/components/schemas/BookingExtras'
        notes:
          type: string
          nullable: true
        vehicle:
          type: object
          properties:
            id:
              type: string
            make:
              type: string
            model:
              type: string
            year:
              type: integer
            images:
              type: array
              items:
                type: string
            category:
              $ref: '#/components/schemas/VehicleCategory'
        user:
          type: object
          properties:
            id:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
        payment:
          type: object
          nullable: true
          properties:
            id:
              type: string
            status:
              $ref: '#/components/schemas/PaymentStatus'
            amount:
              type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateBookingRequest:
      type: object
      required: [vehicleId, startDate, endDate, pickupLocation, dropoffLocation]
      properties:
        vehicleId:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        pickupLocation:
          type: string
          minLength: 1
        dropoffLocation:
          type: string
          minLength: 1
        extras:
          $ref: '#/components/schemas/BookingExtras'
        notes:
          type: string
...
    BookingListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Booking'
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            totalPages:
              type: integer

    # ===== Customer Schemas =====
    Customer:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        emailVerified:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            bookings:
              type: integer
            documents:
              type: integer
...
    CustomerListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Customer'
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            totalPages:
              type: integer

    # ===== Payment Schemas =====
...