# server/src/routes/stats.ts
import { Router } from 'express';
import { z } from 'zod';
import prisma from '../lib/prisma.js';
import { authenticate, staffOnly } from '../middleware/auth.js';

const router = Router();

// GET /api/stats/dashboard - Admin dashboard stats
router.get('/dashboard', authenticate, staffOnly, async (_req, res, next) => {
  try {
    const today = new Date();
    today.setHours(0, 0, 0, 0);

    const tomorrow = new Date(today);
    tomorrow.setDate(tomorrow.getDate() + 1);

    // Get all stats in parallel
    const [
      activeRentals,
      todaysRevenue,
      pendingBookings,
      availableVehicles,
      totalCustomers,
      totalBookings,
      recentBookings,
    ] = await Promise.all([
      // Active rentals count
      prisma.booking.count({
        where: { status: 'ACTIVE' },
      }),

      // Today's revenue
      prisma.payment.aggregate({
        where: {
          status: 'SUCCEEDED',
          createdAt: {
            gte: today,
            lt: tomorrow,
          },
        },
        _sum: { amount: true },
      }),

      // Pending bookings count
      prisma.booking.count({
        where: { status: 'PENDING' },
      }),

      // Available vehicles count
      prisma.vehicle.count({
        where: { status: 'AVAILABLE' },
      }),

      // Total customers
      prisma.user.count({
        where: { role: 'CUSTOMER' },
      }),

      // Total bookings
      prisma.booking.count(),

      // Recent bookings
      prisma.booking.findMany({
        take: 5,
        orderBy: { createdAt: 'desc' },
        include: {
          user: {
            select: { firstName: true, lastName: true, email: true },
          },
          vehicle: {
            select: { make: true, model: true, year: true },
          },
        },
      }),
    ]);

    res.json({
      success: true,
      data: {
        metrics: {
          activeRentals,
          todaysRevenue: Number(todaysRevenue._sum.amount) || 0,
          pendingBookings,
          availableVehicles,
          totalCustomers,
          totalBookings,
        },
        recentBookings,
      },
    });
  } catch (error) {
    next(error);
  }
});

// GET /api/stats/revenue - Revenue over time
router.get('/revenue', authenticate, staffOnly, async (req, res, next) => {
  try {
    const { period = '30' } = z
      .object({
        period: z.enum(['7', '30', '90', '365']).default('30'),
      })
      .parse(req.query);

    const days = parseInt(period);
    const startDate = new Date();
    startDate.setDate(startDate.getDate() - days);

    // Get daily revenue
    const payments = await prisma.payment.findMany({
      where: {
        status: 'SUCCEEDED',
        createdAt: { gte: startDate },
      },
      select: {
        amount: true,
        createdAt: true,
      },
    });

    // Group by date
    const revenueByDate = new Map<string, { revenue: number; count: number }>();

    for (let i = 0; i < days; i++) {
      const date = new Date(startDate);
      date.setDate(date.getDate() + i);
      const dateStr = date.toISOString().split('T')[0];
      revenueByDate.set(dateStr, { revenue: 0, count: 0 });
    }

    payments.forEach((payment) => {
      const dateStr = payment.createdAt.toISOString().split('T')[0];
      const existing = revenueByDate.get(dateStr);
      if (existing) {
        existing.revenue += Number(payment.amount);
        existing.count += 1;
      }
    });

    const data = Array.from(revenueByDate.entries()).map(([date, values]) => ({
      date,
      revenue: values.revenue,
      bookings: values.count,
    }));

    res.json({
      success: true,
      data,
    });
  } catch (error) {
    next(error);
  }
});

// GET /api/stats/fleet - Fleet utilization stats
router.get('/fleet', authenticate, staffOnly, async (_req, res, next) => {
  try {
    const [statusCounts, categoryCounts] = await Promise.all([
      // Vehicles by status
      prisma.vehicle.groupBy({
        by: ['status'],
        _count: { id: true },
      }),

      // Vehicles by category
      prisma.vehicle.groupBy({
        by: ['category'],
        _count: { id: true },
      }),
    ]);

    const totalVehicles = statusCounts.reduce((sum, s) => sum + s._count.id, 0);

    const utilizationRate =
      totalVehicles > 0
        ? ((statusCounts.find((s) => s.status === 'RENTED')?._count.id || 0) /
            totalVehicles) *
          100
        : 0;

    res.json({
      success: true,
      data: {
        totalVehicles,
        utilizationRate: Math.round(utilizationRate * 10) / 10,
        byStatus: statusCounts.map((s) => ({
          status: s.status,
          count: s._count.id,
        })),
        byCategory: categoryCounts.map((c) => ({
          category: c.category,
          count: c._count.id,
        })),
      },
    });
  } catch (error) {
    next(error);
  }
});

// GET /api/stats/bookings - Booking statistics
router.get('/bookings', authenticate, staffOnly, async (req, res, next) => {
  try {
    const { period = '30' } = z
      .object({
        period: z.enum(['7', '30', '90', '365']).default('30'),
      })
      .parse(req.query);

    const days = parseInt(period);
    const startDate = new Date();
    startDate.setDate(startDate.getDate() - days);

    const [statusCounts, bookingsByDay] = await Promise.all([
      // Bookings by status
      prisma.booking.groupBy({
        by: ['status'],
        _count: { id: true },
      }),

      // Bookings over time
      prisma.booking.findMany({
        where: {
          createdAt: { gte: startDate },
        },
        select: {
          createdAt: true,
          status: true,
        },
      }),
    ]);

    // Group bookings by date
    const bookingsByDate = new Map<string, number>();

    for (let i = 0; i < days; i++) {
      const date = new Date(startDate);
      date.setDate(date.getDate() + i);
      const dateStr = date.toISOString().split('T')[0];
      bookingsByDate.set(dateStr, 0);
    }

    bookingsByDay.forEach((booking) => {
      const dateStr = booking.createdAt.toISOString().split('T')[0];
      const existing = bookingsByDate.get(dateStr);
      if (existing !== undefined) {
        bookingsByDate.set(dateStr, existing + 1);
      }
    });

    const trend = Array.from(bookingsByDate.entries()).map(([date, count]) => ({
      date,
      count,
    }));

    res.json({
      success: true,
      data: {
        byStatus: statusCounts.map((s) => ({
          status: s.status,
          count: s._count.id,
        })),
        trend,
      },
    });
  } catch (error) {
    next(error);
  }
});

// GET /api/stats/customers - Customer statistics
router.get('/customers', authenticate, staffOnly, async (req, res, next) => {
  try {
    const { period = '30' } = z
      .object({
        period: z.enum(['7', '30', '90', '365']).default('30'),
      })
      .parse(req.query);

    const days = parseInt(period);
    const startDate = new Date();
    startDate.setDate(startDate.getDate() - days);

    const [totalCustomers, newCustomers, topCustomers] = await Promise.all([
      // Total customers
      prisma.user.count({
        where: { role: 'CUSTOMER' },
      }),

      // New customers in period
      prisma.user.count({
        where: {
          role: 'CUSTOMER',
          createdAt: { gte: startDate },
        },
      }),

      // Top customers by booking count
      prisma.user.findMany({
        where: { role: 'CUSTOMER' },
        take: 10,
        orderBy: {
          bookings: { _count: 'desc' },
        },
        select: {
          id: true,
          firstName: true,
          lastName: true,
          email: true,
          _count: { select: { bookings: true } },
        },
      }),
    ]);

    res.json({
      success: true,
      data: {
        totalCustomers,
        newCustomers,
        topCustomers: topCustomers.map((c) => ({
          id: c.id,
          name: `${c.firstName} ${c.lastName}`,
          email: c.email,
          bookingCount: c._count.bookings,
        })),
      },
    });
  } catch (error) {
    next(error);
  }
});

export default router;